blockchain: add tests for BlockStore
Add tests to test store, to the fullest reasonable extent for
paths that can taken by input arguments altering internal behavior.

The test data can be generated by running the code at
https://gist.github.com/odeke-em/9ffac2b5df44595fad7084ece4c9bd98
which I haven't included in the test because it makes the noisy
and not too easy to grok since that's more code surface.
I've included the gist instead for reference.